Separate date mixin tests from other model tests
