+++ geniza/corpus/tests/test_corpus_models.py
@@ -8,7 +8,6 @@ from django.contrib.admin.models import ADDITION, CHANGE, LogEntry
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
-from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 from django.utils import timezone
 from django.utils.safestring import SafeString
@@ -17,7 +16,6 @@ from djiffy.models import Canvas, IIIFException, IIIFImage, Manifest
 from modeltranslation.manager import MultilingualQuerySet
 from piffle.presentation import IIIFException as piffle_IIIFException
 
-from geniza.corpus.dates import Calendar
 from geniza.corpus.models import (
     Collection,
     Document,
@@ -439,52 +437,6 @@ class TestDocument:
 
         unsaved_doc = Document()
         assert str(unsaved_doc) == "?? (PGPID ??)"
-
-    def test_clean(self):
-        doc = Document()
-        # no dates; no error
-        doc.clean()
-
-        # original date but no calendar — error
-        doc.doc_date_original = "480"
-        with pytest.raises(ValidationError):
-            doc.clean()
-
-        # calendar but no date — error
-        doc.doc_date_original = ""
-        doc.doc_date_calendar = Calendar.HIJRI
-        with pytest.raises(ValidationError):
-            doc.clean()
-
-        # both — no error
-        doc.doc_date_original = "350"
-        doc.clean()
-
-    def test_original_date(self):
-        """Should display the historical document date with its calendar name"""
-        doc = Document.objects.create(
-            doc_date_original="507", doc_date_calendar=Calendar.HIJRI
-        )
-        assert doc.original_date == "507 Hijrī"
-        # with no calendar, just display the date
-        doc.doc_date_calendar = ""
-        assert doc.original_date == "507"
-
-    def test_document_date(self):
-        """Should combine historical and converted dates"""
-        doc = Document.objects.create(
-            doc_date_original="507",
-            doc_date_calendar=Calendar.HIJRI,
-        )
-        # should just use the original_date method
-        assert doc.document_date == doc.original_date
-        # should wrap standard date in parentheses and add CE
-        doc.doc_date_standard = "1113/14"
-        assert doc.document_date == "507 Hijrī (1113/14 CE)"
-        # should return standard date only, no parentheses
-        doc.doc_date_original = ""
-        doc.doc_date_calendar = ""
-        assert doc.document_date == "1113/14 CE"
 
     def test_collection(self):
         # T-S 8J22.21 + T-S NS J193
